// src/types/routes.ts
export enum ROUTES {
  CLIENT = "/",
  CLIENT_DEALS = "/my-deals",
  CLIENT_CREATE_ORDER = "/create-order",
  CLIENT_SEARCH_CANDIDATES = "/search-candidates",

  FREELANCER = "/freelancer",
  FREELANCER_JOBS = "/freelancer/my-jobs",
  FREELANCER_JOB_SEARCH = "/freelancer/job-search",
  FREELANCER_WITHDRAW_FUNDS = "/freelancer/withdraw-funds",

  SUPPORT = "/config/support",
  SETTINGS = "/config/settings",
}